Put back the line of code that writes block data to disk.
Without this line blocks are not written to the data file.
The block index is still written. When it is read back the client
will crash because it cannot read the corresponding data.

Anyone who ran the buggy code without this fix and received a
block from a peer will need to redownload the blockchain.